feat(psnr_bitrate_parser): added reusability features
Made it so psnr_bitrate_parser now creates and saves unique txt files for each sequence encoded, and checks if said txt file already exists in target dir. If it doesn't, the script proceeds to encode the sequence, but if it does then the script does nothing as the result would be the same, thus saving time. Json config file was also changed and new optional settings for both vvenc and VTM were added in case more options need to be passed onto the terminal.